migrated other classes, optimization functions
