+++ burnman/classes/planet.py
@@ -6,7 +6,7 @@ from __future__ import print_function
 
 import numpy as np
 import warnings
-from .classes.material import material_property
+from .material import material_property
 
 
 class Planet(object):
