+++ burnman/classes/seismic.py
@@ -10,8 +10,8 @@ import numpy as np
 import warnings
 import scipy.integrate
 
-from .tools.misc import read_table
-from . import constants
+from ..tools.misc import read_table
+from .. import constants
 
 
 class Seismic1DModel(object):
