+++ burnman/optimize/__init__.py
@@ -1,0 +1,15 @@
+# This file is part of BurnMan - a thermoelastic and thermodynamic toolkit for
+# the Earth and Planetary Sciences
+# Copyright (C) 2012 - 2021 by the BurnMan team, released under the GNU
+# GPL v2 or later.
+
+"""
+Optimization functions
+(fitting and solving)
+"""
+
+from . import composition_fitting
+from . import eos_fitting
+from . import linear_fitting
+from . import nonlinear_fitting
+from . import nonlinear_solvers
