+++ burnman/optimize/eos_fitting.py
@@ -1,4 +1,5 @@
-# This file is part of BurnMan - a thermoelastic and thermodynamic toolkit for the Earth and Planetary Sciences
+# This file is part of BurnMan - a thermoelastic and thermodynamic toolkit for
+# the Earth and Planetary Sciences
 # Copyright (C) 2012 - 2015 by the BurnMan team, released under the GNU
 # GPL v2 or later.
 
@@ -8,8 +9,8 @@ from __future__ import print_function
 import numpy as np
 
 from . import nonlinear_fitting
-from .tools.misc import flatten
-from .tools.math import unit_normalize
+from ..tools.misc import flatten
+from ..tools.math import unit_normalize
 
 
 def fit_PTp_data(mineral, fit_params, flags, data, data_covariances=[], mle_tolerances=[], param_tolerance=1.e-5, max_lm_iterations=50, verbose=True):
