+++ burnman/optimize/nonlinear_solvers.py
