+++ burnman/tools/equilibration.py
@@ -10,7 +10,7 @@ import numpy as np
 from scipy.linalg import lu_factor, lu_solve
 from collections import namedtuple
 
-from ..nonlinear_solvers import damped_newton_solve
+from ..optimize.nonlinear_solvers import damped_newton_solve
 from ..classes.solidsolution import SolidSolution
