+++ docs/api.rst
@@ -10,6 +10,7 @@ Autogenerated Full API
    polytopes
    averaging
    geotherms
+   planets
    thermodynamics
    seismic
    mineral_database
