+++ docs/compositions.rst
@@ -10,12 +10,12 @@ Base class
 Utility functions
 -----------------
 
-.. autofunction:: burnman.composition.file_to_composition_list
+.. autofunction:: burnman.classes.composition.file_to_composition_list
 
 
 Fitting functions
 -----------------
 
-.. autofunction:: burnman.composition_fitting.fit_composition_to_solution
+.. autofunction:: burnman.optimize.composition_fitting.fit_composition_to_solution
 
-.. autofunction:: burnman.composition_fitting.fit_phase_proportions_to_bulk_composition
+.. autofunction:: burnman.optimize.composition_fitting.fit_phase_proportions_to_bulk_composition
