+++ docs/planets.rst
@@ -1,0 +1,13 @@
+Planets and Layers
+==================
+
+Layer
+-----
+
+.. autoclass:: burnman.Layer
+
+
+Planet
+------
+
+.. autoclass:: burnman.Planet
