+++ docs/seismic.rst
@@ -4,32 +4,32 @@ Seismic
 
 Base class for all seismic models
 ---------------------------------
-.. autoclass:: burnman.seismic.Seismic1DModel
+.. autoclass:: burnman.classes.seismic.Seismic1DModel
 
 
 Class for 1D Models
 -------------------
-.. autoclass:: burnman.seismic.SeismicTable
+.. autoclass:: burnman.classes.seismic.SeismicTable
 
 
 
 Models currently implemented
 ----------------------------
 
-.. autoclass:: burnman.seismic.PREM
+.. autoclass:: burnman.classes.seismic.PREM
 
-.. autoclass:: burnman.seismic.Slow
+.. autoclass:: burnman.classes.seismic.Slow
 
-.. autoclass:: burnman.seismic.Fast
+.. autoclass:: burnman.classes.seismic.Fast
 
-.. autoclass:: burnman.seismic.STW105
+.. autoclass:: burnman.classes.seismic.STW105
 
-.. autoclass:: burnman.seismic.IASP91
+.. autoclass:: burnman.classes.seismic.IASP91
 
-.. autoclass:: burnman.seismic.AK135
+.. autoclass:: burnman.classes.seismic.AK135
 
 
 Attenuation Correction
 -----------------------
 
-.. autofunction:: burnman.seismic.attenuation_correction
+.. autofunction:: burnman.classes.seismic.attenuation_correction
