+++ tests/test_solvers.py
@@ -3,7 +3,7 @@ from util import BurnManTest
 import numpy as np
 
 import burnman_path
-from burnman.nonlinear_solvers import damped_newton_solve
+from burnman.optimize.nonlinear_solvers import damped_newton_solve
 
 assert burnman_path  # silence pyflakes warning
